convert the CheckInService to an IntentService and have it scheduled via AlarmManager, to run every 15 mins
